Extract `FeedImageDataLoaderSpy` into a shared helper folder to remove duplication
